backend/pipelines/open-banking-schedule-news/requirements.txt
newsapi-python
numpy
pandas
requests
datetime